requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Faker==8.1.2
Flask==1.1.2
gunicorn==20.0.4
html2text==2020.1.16
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
peewee==3.14.4
pycurl==7.43.0.6
pydantic==1.8.1
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.0
requests==2.25.1
six==1.15.0
text-unidecode==1.3
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wptools==0.4.17
